% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_sha1}
\alias{compare_sha1}
\title{compare 2 vectors}
\usage{
compare_sha1(a, b)
}
\arguments{
\item{a}{first vector of repo names}

\item{b}{second vector of repo names}
}
\description{
compare 2 vectors
}
