% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{local_gh}
\alias{local_gh}
\title{return data.frame of local github packages}
\usage{
local_gh(lib = NULL)
}
\arguments{
\item{lib}{path to local lib}
}
\value{
a named character
}
\description{
return data.frame of local github packages
}
