% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_version}
\alias{local_version}
\title{modified from r-lib/sessioninfo licence GPL/2}
\usage{
local_version(desc)
}
\arguments{
\item{desc}{pkg description}
}
\description{
modified from r-lib/sessioninfo licence GPL/2
}
