% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remote_version}
\alias{remote_version}
\title{build remote version, ref and short sha1}
\usage{
remote_version(ref, sha)
}
\arguments{
\item{ref}{vector of repo ref}

\item{sha}{vector of sha1sum}
}
\description{
build remote version, ref and short sha1
}
