% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{upnews}
\alias{upnews}
\title{fetch news from remote github repo}
\usage{
upnews(lib = NULL, debug = FALSE)
}
\arguments{
\item{lib}{path to local lib}

\item{debug}{boolean for displaying installed gh}
}
\value{
a data.frame
}
\description{
fetch news from remote github repo
}
